Conditions to not set State if no data in db
The app would throw an error when opened on a day on which there's no data yet at the rtdb (realtime database).
To fix this I added conditions to not set states if the data from database is null/undefined but only when getting data from a node that changes daily.